Temp fix for #252
When a search request is made, all the fields are returned... this includes all markdown in the `content` field. To overcome this issue temporarily I'm excluding this field from the query, but this may cause issues if a user doesn't have a `content` field in the collection. In which case we can check against the config file.